// apps/demos/Demos/Toolbar/Adaptability/React/types.ts
export type FontSize = {
  size: number;
  text: string;
};

export type LineHeight = {
  lineHeight: number;
  text: string;
};

export type FontFamily = {
  text: string;
};

export type Heading = {
  text: string;
};

export type FontStyle = {
  icon: string;
  hint: string;
};

export type TextAlignItem = {
  icon: string;
  alignment: string;
  hint: string;
  text?: string;
};

export type ListType = {
  icon: string;
  alignment: string;
  hint: string;
};